Fix set and disk_set to not crash when no flags are supported
Loop labels and file images support no flags.  set and disk_set
would prompt for a flag and accept any string since the list of
flags was empty, then fail to look up an actual flag value, then
throw an exception with a null string for the name of the flag,
which would bug.